fixes for cross-compile - under Arduino SDK max is a macro, so arguments can be different. Under regular gcc max is a template function expecting same types for a/b - the 127 constant is by default an int, not an int8_t. - config.h needs to come from the include path and not look in the current directory first

+++ TempSensor.cpp
@@ -43,7 +43,7 @@ void TempSensor::update()
 	temperature temp;
 	if (!_sensor || (temp=_sensor->read())==TEMP_SENSOR_DISCONNECTED) {		
 		failedReadCount++;		
-		failedReadCount = min(failedReadCount,127);	// limit
+		failedReadCount = min(failedReadCount,int8_t(127));	// limit
 		return;
 	}
